scripts/requirements-whisper.txt
# Local transcription only (pnpm voices:transcribe runs this via Python).
# Install into a venv: pip install -r scripts/requirements-whisper.txt
openai-whisper>=20231117
